+++ src/nespresso/db/services/user.py
@@ -51,7 +51,15 @@ class UserService:
             column=TgUser.chat_id,
         )
 
-        return result if result else []
+        return result
+
+    async def GetActiveMatchingTgUsersChatId(self) -> list[int]:
+        result = await self.GetTgUsersOnCondition(
+            condition=TgUser.active_matching,
+            column=TgUser.chat_id,
+        )
+
+        return result
 
     # ----- Update -----