+++ src/nespresso/bot/lib/chat/username.py
@@ -28,3 +28,9 @@ async def GetTgUsername(chat_id: int) -> str | None:
     except Exception as e:
         logging.warning(f"Failed to get chat info for chat_id={chat_id}: {e}")
         return None
+
+
+async def GetChatUserLoggingPart(chat_id: int) -> str:
+    username = await GetTgUsername(chat_id) or "-/-"
+
+    return f"chat_id={chat_id:<10} ({username + ")":<25}"