+++ CHANGELOG.md
@@ -32,6 +32,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Matching Framework
 
+### Updated
+
+- DB refactoring
+
 ## [0.1.0] - 2025-06-10 - Vlad
 
 ### Added